requirements.txt
qrcode
barcode
python-barcode
pillow